+++ Makefile
@@ -3,7 +3,7 @@ CC			=	cc
 LDFLAGS		=	-L${HOME}/.brew/opt/readline/lib -lreadline
 CFALG		=	-I${HOME}/.brew/opt/readline/include
 SOURCE		=	main.c src/read_input.c src/sig_handler.c src/term.c src/utils1.c src/init_minishell.c \
-				src/parse_input.c src/split.c
+				src/parse_input.c src/split.c src/split_utils.c
 HEADER		=	minishell.h
 LIBRARY 	=	libft/libft.a
