+++ main.c
@@ -6,7 +6,7 @@
 /*   By: shikim <shikim@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/11 16:57:01 by shikim            #+#    #+#             */
-/*   Updated: 2023/07/13 19:48:19 by shikim           ###   ########.fr       */
+/*   Updated: 2023/07/14 21:42:16 by shikim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
